DEV-2217 Update scikit-allel dependency
The pipeline5 imager script recently stopped working with the version of
this dependency that was pinned in the requirements (==1.2.1). It wasn't
entirely clear what was happening but I believe there were conflicts
with some of the other versions specified in the requirements.txt.

I arrived at the new version by removing the version from the
requirements and reinstalling in a new virtual environment, essentially
allowing `pip` to select the new version. There may be better ways to do
this but I'm no expert and this at least allows the image to be built.
Of course further testing is required, either standalone or as part of
the pipeline 5.24 release.

I also added in the `setuptools` dependency that `pip` decided on as it
seems it was previously being added implicitly.