Stop using @classmethod so much
It did successfully indicate that a class using it a lot
does not make use of internal state, but it's
a more advanced Python feature
then is strictly required for PEACH,
which is an additional barrier to understanding the code
for people with limited experience with Python.
Because it is a decorator it also messed with IDE
assistance in refactoring etc.