DEV-2683: Set number of PAVE_TI entries to read to 1000
This is done to get around the fact that scikit-allel
only reads 1 PAVE_TI annotation if "Number=." for this field
in the VCF header, even though this indicates the number
of entries for this field can vary. Maximum number of
transcripts per gene is 241 in our current gene panel,
so 1000 seems like a reasonable choice. 10000 would slow
PEACH down significantly. It's fine to read more
entries than exist, since then it pads with empty strings.