DEV-2837: Remove comparison of VCF and JSON gene names
UGT1A1 is entered in driver gene panel as UGT1A8 because of
the ensembl ID swap for these genes in 37 wrt 38. This problem is
unavoidable for the moment, so a mismatch between VCF and JSON gene
names is unavoidable for the moment as well. This comparison is the
only use of the VCF gene name, so this commit removes the reading
of the VCF gene name all together.